Broke Settings components into smaller files for readability - Created InputForm.js and QueryComponent.js These are imported into Settings.js
