+++ client/src/components/Settings/Settings.js
@@ -1,18 +1,9 @@
 import React, { Component } from 'react'
-import { Query, Mutation } from 'react-apollo'
-import PropTypes from 'prop-types'
+import { Mutation } from 'react-apollo'
 import gql from 'graphql-tag'
-import { InputGroup, InputGroupAddon, Input, Button } from 'reactstrap'
+import { Button } from 'reactstrap'
+import QueryComponent from './QueryComponent'
 import './Settings.css'
-
-// GraphQL Query for current user information
-const getCurrentInformation = gql`
-  {
-    teacher(encJwt: "${window.localStorage.getItem('token')}") {
-      TeacherName
-      TeacherEmail
-    }
-  }`
 
 // GraphQL Mutation to update user information
 const updateInformation = gql`
@@ -25,76 +16,6 @@ const updateInformation = gql`
       jwtString
     }
   }`
-
-const InputForm = (props) => {
-  return (
-    <div>
-      <InputGroup>
-        <InputGroupAddon addonType='prepend'>Name:</InputGroupAddon>
-        <Input placeholder={props.data.teacher ? props.data.teacher[0].TeacherName : 'Loading...'} type='text' name='name' value={props.name} onChange={props.handleInputChange} />
-      </InputGroup>
-      <br />
-      <InputGroup>
-        <InputGroupAddon addonType='prepend'>Email:</InputGroupAddon>
-        <Input placeholder={props.data.teacher ? props.data.teacher[0].TeacherEmail : 'Loading...'} type='email' name='email' value={props.email} onChange={props.handleInputChange} />
-      </InputGroup>
-      <br />
-      <InputGroup>
-        <InputGroupAddon addonType='prepend'>Old Password: </InputGroupAddon>
-        <Input type='password' name='oldPassword' value={props.oldPassword} onChange={props.handleInputChange} />
-      </InputGroup>
-      <br />
-      <InputGroup>
-        <InputGroupAddon addonType='prepend'>New Password: </InputGroupAddon>
-        <Input type='password' name='newPassword' value={props.newPassword} onChange={props.handleInputChange} />
-      </InputGroup>
-    </div>
-  )
-}
-
-InputForm.propTypes = {
-  data: PropTypes.object,
-  name: PropTypes.string,
-  email: PropTypes.string,
-  handleInputChange: PropTypes.func,
-  oldPassword: PropTypes.string,
-  newPassword: PropTypes.string
-}
-
-class QueryComponent extends Component {
-  constructor (props) {
-    super()
-    this.state = {
-      name: props.name,
-      email: props.email,
-      oldPassword: props.oldPassword,
-      newPassword: props.newPassword
-    }
-  }
-
-  render () {
-    return (
-      <Query query={getCurrentInformation}>
-        {({ loading, error, data }) => {
-          if (loading) return 'Loading...'
-          if (error) return `Error: ${error.message}`
-
-          return (
-            <InputForm data={data} handleInputChange={this.props.handleInputChange} />
-          )
-        }}
-      </Query>
-    )
-  }
-}
-
-QueryComponent.propTypes = {
-  name: PropTypes.string,
-  email: PropTypes.string,
-  oldPassword: PropTypes.string,
-  newPassword: PropTypes.string,
-  handleInputChange: PropTypes.func
-}
 
 class Settings extends Component {
   constructor () {
